requirements.txt
django==4.2
pillow
django-admin-tailwind
setuptools
django-paypal==2.1